Move build-option methods from commands into builder
Initial work added the build-options into commands/build.go under preRunBuild.
In testing this was found to mean that --lang was having to be added to the
CLI even though it was already specified in the yaml.

This work has refactored the build-options work to move it into the builder package
at which point the language, whether from the CLI or yaml, is known and so negates the
need for the user to 'double supply'.

The basic flow is bring together all desired build-options, determine that they are all
valid for the current language, convert the build-options to a list of packages, deduplicate
the packages, convert to a build-arg of ADDITIONAL_PACKAGES= before appending any build-args
od the same name to the set.

Signed-off-by: rgee0 <richard@technologee.co.uk>